fix: relax assumptions about tempus api data types
this commit changes all unsigned integers in
crate::models and crate::client to their signed
counterparts. crate::db is updated to remove the
associated casts since they're not longer needed.
since postgres integer types are signed anyway,
reducing the assumptions about what tempus returns
eliminates a whole class of errors around edge
cases where tempus returns unusual data, such as
model::DemoInfo::start_time being negative in just
one run out of millions. in addition, a large test
file is added to make sure this edge case is handled.

+++ src/db.rs
@@ -23,7 +23,7 @@ pub async fn upsert_maps(pool: &PgPool, maps: &[Map]) -> anyhow::Result<()> {
                 video_demoman = EXCLUDED.video_demoman,
                 fetched_at = EXCLUDED.fetched_at
             "#,
-            map.id as i32,
+            map.id,
             &map.name,
             map.tier_info.soldier as i16,
             map.tier_info.demoman as i16,
@@ -51,7 +51,7 @@ pub async fn upsert_users(pool: &PgPool, records: &[Record]) -> anyhow::Result<(
                 steamid = EXCLUDED.steamid,
                 name = EXCLUDED.name
             "#,
-            rec.user_id as i32,
+            rec.user_id,
             &rec.steamid,
             &rec.name,
         )
@@ -74,9 +74,9 @@ pub async fn upsert_records(pool: &PgPool, map_id: i32, records: &[Record]) -> a
                 date = EXCLUDED.date,
                 fetched_at = EXCLUDED.fetched_at
             "#,
-            rec.id as i32,
+            rec.id,
             map_id,
-            rec.user_id as i32,
+            rec.user_id,
             rec.class as i16,
             rec.duration,
             rec.date,
